first migration into DNAEdge
